net-dns/dnsmasq: Add support for multiple instances
To make sure this works nicely few related things need to be changed.  Due to
the fact that different instances may run under different users, the user and
group now need to be specified separately.  That is due to them being used in
the start_pre() function when checking permissions for the leases file.

The changes are only in the confd and initd files, the ebuild is unchanged
otherwise (Hence only the release bump).

Closes: https://bugs.gentoo.org/282397
Package-Manager: Portage-2.3.28, Repoman-2.3.9