net-p2p/retroshare update development ebuild
Development version of retroshare expose qmake flags that permit easier
build customization without crufty patching, improve development ebuild
to take advantage of that and have a cleaner ebuild.
Removed VoIP and feedreader plugins are not maintained.
For future stable versions this ebuild should be used as reference and
not the older ones.